Add all steps (for real). NEXT: Fix bug with save button
