--- src/components/sequences/steps/send_message_step.tsx
@@ -1,96 +1,0 @@
-import * as React from "react";
-import { changeStep, removeStep, pushStep } from "../sequence_actions";
-import { assign } from "lodash";
-import { Step, Sequence } from "../interfaces";
-
-interface CopyParams {
-  dispatch: Function;
-  step: Step;
-  index: number;
-}
-
-function copy({dispatch, step, index}: CopyParams) {
-  let copy = assign<{}, Step>({}, step);
-  dispatch(pushStep(copy, (index + 1)));
-};
-
-interface RemoveParams {
-  index: number;
-  dispatch: Function;
-}
-
-function remove({dispatch, index}: RemoveParams) {
-  dispatch(removeStep(index));
-}
-
-interface UpdateStepParams {
-  dispatch: Function;
-  step: Step;
-  index: number;
-  field: string; // "x"|"y"|"z"|"speed";
-}
-
-let updateStep = function ({ dispatch,
-                             step,
-                             index,
-                             field }: UpdateStepParams) {
-  return (e) => {
-    let update = assign<{}, Step>({}, step);
-    update.command[field] = e.target.value;
-    let action = changeStep(index, update);
-    dispatch(action);
-  };
-};
-
-interface IStepInputBox {
-  step: Step;
-  field: string;
-  dispatch: Function;
-  index: number;
-}
-function StepInputBox({step, field, dispatch, index}: IStepInputBox) {
-  return <input type="text"
-                value={ step.command[field] }
-                onChange={ updateStep({dispatch, step, index, field}) } />;
-}
-
-interface SendMessageStepParams {
-  dispatch: Function;
-  step: Step;
-  index: number;
-}
-export function SendMessageStep({dispatch,
-                                    step,
-                                    index}: SendMessageStepParams) {
-    return( <div>
-              <div className="step-wrapper">
-                <div className="row">
-                  <div className="col-sm-12">
-                    <div className="step-header send-message-step">
-                      <input className="step-label" placeholder="Send Message"/>
-                      <i className="fa fa-arrows-v step-control" />
-                      <i className="fa fa-clone step-control"
-                         onClick={ () => copy({dispatch, step, index}) } />
-                      <i className="fa fa-trash step-control"
-                         onClick={ () => remove({dispatch, index}) } />
-                    </div>
-                  </div>
-                </div>
-                <div className="row">
-                  <div className="col-sm-12">
-                    <div className="step-content send-message-step">
-                      <div className="row">
-                        <div className="col-xs-12">
-                          <label>Message</label>
-                          <StepInputBox dispatch={dispatch}
-                                        step={step}
-                                        index={index}
-                                        field="message"/>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div> );
-}
